Extract Customize Responses state derivation into shared AIChat package
